add funt that player is move

+++ include/cub3d.h
@@ -6,7 +6,7 @@
 /*   By: hsano <hsano@student.42tokyo.jp>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/09 13:18:21 by hsano             #+#    #+#             */
-/*   Updated: 2023/01/07 08:40:33 by hsano            ###   ########.fr       */
+/*   Updated: 2023/01/07 09:06:43 by hsano            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -108,7 +108,6 @@ typedef enum e_wall_dir
 
 typedef struct s_ray
 {
-	//int		x;
 	t_wall_dir	wall_dir;
 	t_cub3d_type	begin_angle;
 	t_cub3d_type	last_angle;
@@ -116,20 +115,12 @@ typedef struct s_ray
 	t_cub3d_type	stop_angle;
 	t_cub3d_type	begin_base_len;
 	t_cub3d_type	last_base_len;
-	//t_cub3d_type	img_offset_begin;
-	//t_cub3d_type	img_offset_last;
 	t_cub3d_type	max_len;
 	int		wall_pixel;
-	//int		begin_x;
-	//t_point		distance;
 	t_point		map_point;
 	t_point		begin_distance;
-	//t_point		base_distance;
 	t_point		last_distance;
 	t_image		*wall_img;
-	//t_cub3d_type	ratio;
-	//int		next_i;
-	//int		is_front_wall;
 	int		is_adjacent_wall;
 }	t_ray;
 
@@ -141,8 +132,8 @@ typedef struct s_player
 	//int	y; // > 0 && < MAP_SPACE
 	t_point	map;
 	t_point	mass;
-	int	world_x; // map_x * MAP_SPACE + x
-	int	world_y; // map_y * MAP_SPACE + y
+	size_t	world_x; // map_x * MAP_SPACE + x
+	size_t	world_y; // map_y * MAP_SPACE + y
 	t_angle	dir; //north:0 west:90 south:180 east:270
 }	t_player;
 
@@ -155,7 +146,6 @@ typedef struct s_cub3d
 	t_wall_imgs	*walls;
 	t_wall_imgs	*trans_walls;
 	t_player	*player;
-	//t_ray		rays[WIN_WIDTH * 360 / FOV];
 	t_angle	angles[WIN_WIDTH * 360 / FOV];
 	int		lock;
 }	t_cub3d;
